Fix shill unittest build problem for gcc 4.7.
gcc 4.7 is more strict about type conversion in initilization list,
for example -

const char apple[] = {0xff, 0xfe};

The above compiles ok prior to gcc 4.7, but fails on gcc 4.7 with
-Wall -Werror. The reason is that 0xff is out of bound for (signed)
char type.

TEST=Built/tested using gcc 4.7
BUG=None

Reviewed-on: https://gerrit.chromium.org/gerrit/31908
Commit-Ready: Han Shen <shenhan@chromium.org>
Tested-by: Han Shen <shenhan@chromium.org>

+++ shill/key_value_store_unittest.cc
@@ -27,8 +27,13 @@ TEST_F(KeyValueStoreTest, Bool) {
   EXPECT_EQ(kDefaultValue, store_.LookupBool(kKey, kDefaultValue));
   store_.SetBool(kKey, kValue);
   EXPECT_TRUE(store_.ContainsBool(kKey));
-  EXPECT_EQ(kValue, store_.LookupBool(kKey, kDefaultValue));
-  EXPECT_EQ(kValue, store_.GetBool(kKey));
+  // TODO: investigate if a newer version of gtest handles EXPECT_EQ for bools
+  // in a manner that gcc 4.7 is happy with. (Inproper conversion from "false"
+  // to "NULL".)
+  EXPECT_EQ(static_cast<int>(kValue),
+            static_cast<int>(store_.LookupBool(kKey, kDefaultValue)));
+  EXPECT_EQ(static_cast<int>(kValue),
+            static_cast<int>(store_.GetBool(kKey)));
 }
 
 TEST_F(KeyValueStoreTest, Int) {
